revert file rearrangement, cythonize main.py
