+++ pydantic/__init__.py
@@ -5,9 +5,8 @@ from .env_settings import BaseSettings
 from .error_wrappers import ValidationError
 from .errors import *
 from .fields import Required
-from .main import BaseConfig, BaseModel, Extra, create_model
-from .parse import compiled, parse_model
-from .parse_files import Protocol
+from .main import BaseConfig, BaseModel, Extra, compiled, create_model, validate_model
+from .parse import Protocol
 from .schema import Schema
 from .types import *
 from .version import VERSION
