+++ setup.cfg
@@ -5,7 +5,7 @@ filterwarnings = error
 
 [flake8]
 max-line-length = 120
-max-complexity = 12
+max-complexity = 14
 
 [bdist_wheel]
 python-tag = py36.py37.py38
